added the assetic settings into config.yml. technically they can be added to congif_prod.yml so that the css and js files are only minified when running in production. add that later.
i have only used assetic on the frontpage so far. i was unable to use it on common.js and I'm not sure why. it must be throwing an error but I don't know how to get an error log back from assetic.
I also wasn't able to use it on main.css, the reason being that the google fonts aren't recognized in the minified file that is generated by assetic. I also don't know why that is.

So I left those two files alone, the rest I minified.