fixed gui bug that wasn't passing the event ID correctly.
also added logic such that if the event is tied to a purchase, the purchase is removed.

commented out the part that emails the user to let them know that their purchase has been removed. I need to find another place to call the mail service since it cannot be called from a repo